+++ NEWS.md
@@ -4,6 +4,7 @@ Released: 2025-02-00
 * Add code generation for the openMOLE platform.
 * Add support for reproducibility, using the doRNG package.
 * Plots for object count and fragment count are displayed, if relevant, in simulation results.
+* The choice of the clustering method in the 'Measurements' tab is now possible.
 
 # archeofrag.gui 0.1.0
 Released: 2024-12-19